[4/8] 1.Done -목록에 최신글 5개 불러와서 링크걸기 -세션에 유저의 타입도 저장해서 타입이 sns면 비밀번호 재설정 안보이게 	-> 로그아웃 코드 변경 -로그인이 필요한 페이지의 경우 로그인 하도록 리다이렉트 	-> 로그인 한 후 접근하려던 페이지로 접근 가능 -댓글 비밀글 설정 	-> 댓글 작성자와 글 작성자만 볼 수 있도록
2.todo
-계층형 게시판
-지도 api? 등 여러 api

+++ CSH/src/main/java/com/gmail/hyunyful/BoardController.java
@@ -32,13 +32,34 @@ public class BoardController {
 	
 	//게시글 작성 페이지 가기
 	@RequestMapping(value="/board/write",method=RequestMethod.GET)
-	public String writePage() {
+	public String writePage(HttpServletRequest request) {
+		//로그인이 안되어 있으면
+		if(request.getSession().getAttribute("userEmail") == null) {
+			//현재의 uri를 세션에 저장(/board/wirte)
+			//request.getRequestURI() 는 포트번호 이후의 경로를 반환 (String 형) ex./board/write
+			//request.getRequestURL() 은 http부터 끝까지의 경로를 반환 (StringBuffer 형) ex.http://localhost:8080/board/write 
+			String uri = request.getRequestURI();
+			request.getSession().setAttribute("uri",uri); 
+			//로그인 페이지가 없으므로 메인페이지로 리다이렉트
+			return "redirect:/";
+		}
+		
 		return "/board/write";
 	}
 	
 	//게시글 작성 
 	@RequestMapping(value="/board/write",method=RequestMethod.POST)
 	public String write(HttpServletRequest request, RedirectAttributes attr) {
+		//로그인이 안되어 있으면
+		if(request.getSession().getAttribute("userEmail") == null) {
+			//현재의 uri를 세션에 저장(/board/wirte)
+			//request.getRequestURI() 는 포트번호 이후의 경로를 반환 (String 형) ex./board/write
+			//request.getRequestURL() 은 http부터 끝까지의 경로를 반환 (StringBuffer 형) ex.http://localhost:8080/board/write 
+			String uri = request.getRequestURI();
+			request.getSession().setAttribute("uri",uri); 
+			//로그인 페이지가 없으므로 메인페이지로 리다이렉트
+			return "redirect:/";
+		}
 		
 		boolean result = service.write(request);
 		attr.addFlashAttribute("write",result);
@@ -48,7 +69,18 @@ public class BoardController {
 	
 	//내가 쓴 글 모아보기
 	@RequestMapping(value="/board/myList/{email:.+}")
-	public String myList(@PathVariable String email, Model model) {
+	public String myList(@PathVariable String email, Model model, HttpServletRequest request) {
+		//로그인이 안되어 있으면
+		if(request.getSession().getAttribute("userEmail") == null) {
+			//현재의 uri를 세션에 저장(/board/wirte)
+			//request.getRequestURI() 는 포트번호 이후의 경로를 반환 (String 형) ex./board/write
+			//request.getRequestURL() 은 http부터 끝까지의 경로를 반환 (StringBuffer 형) ex.http://localhost:8080/board/write 
+			String uri = request.getRequestURI();
+			request.getSession().setAttribute("uri",uri); 
+			//로그인 페이지가 없으므로 메인페이지로 리다이렉트
+			return "redirect:/";
+		}
+		
 		List<Board> list = service.mine(email);
 		model.addAttribute("myList",list);
 		
@@ -67,7 +99,18 @@ public class BoardController {
 	
 	//게시글 상세보기
 	@RequestMapping(value="/board/detail/{bno}")
-	public String detail(@PathVariable int bno, Model model) {
+	public String detail(@PathVariable int bno, Model model, HttpServletRequest request) {
+		//로그인이 안되어 있으면
+		if(request.getSession().getAttribute("userEmail") == null) {
+			//현재의 uri를 세션에 저장(/board/wirte)
+			//request.getRequestURI() 는 포트번호 이후의 경로를 반환 (String 형) ex./board/write
+			//request.getRequestURL() 은 http부터 끝까지의 경로를 반환 (StringBuffer 형) ex.http://localhost:8080/board/write 
+			String uri = request.getRequestURI();
+			request.getSession().setAttribute("uri",uri); 
+			//로그인 페이지가 없으므로 메인페이지로 리다이렉트
+			return "redirect:/";
+		}
+		
 		Board board = service.detail(bno);
 		model.addAttribute("detail",board);
 		
@@ -76,7 +119,18 @@ public class BoardController {
 	
 	//게시글 삭제
 	@RequestMapping(value="/board/delete/{bno}")
-	public String delete(@PathVariable int bno,RedirectAttributes attr) {
+	public String delete(@PathVariable int bno,RedirectAttributes attr, HttpServletRequest request) {
+		//로그인이 안되어 있으면
+		if(request.getSession().getAttribute("userEmail") == null) {
+			//현재의 uri를 세션에 저장(/board/wirte)
+			//request.getRequestURI() 는 포트번호 이후의 경로를 반환 (String 형) ex./board/write
+			//request.getRequestURL() 은 http부터 끝까지의 경로를 반환 (StringBuffer 형) ex.http://localhost:8080/board/write 
+			String uri = request.getRequestURI();
+			request.getSession().setAttribute("uri",uri); 
+			//로그인 페이지가 없으므로 메인페이지로 리다이렉트
+			return "redirect:/";
+		}
+		
 		boolean result = service.delete(bno);
 		//실패한 경우
 		if(result == false) {
@@ -89,7 +143,18 @@ public class BoardController {
 	
 	//게시글 수정 페이지 가기
 	@RequestMapping(value="/board/update/{bno}",method=RequestMethod.GET)
-	public String update(@PathVariable int bno,Model model) {
+	public String update(@PathVariable int bno,Model model, HttpServletRequest request) {
+		//로그인이 안되어 있으면
+		if(request.getSession().getAttribute("userEmail") == null) {
+			//현재의 uri를 세션에 저장(/board/wirte)
+			//request.getRequestURI() 는 포트번호 이후의 경로를 반환 (String 형) ex./board/write
+			//request.getRequestURL() 은 http부터 끝까지의 경로를 반환 (StringBuffer 형) ex.http://localhost:8080/board/write 
+			String uri = request.getRequestURI();
+			request.getSession().setAttribute("uri",uri); 
+			//로그인 페이지가 없으므로 메인페이지로 리다이렉트
+			return "redirect:/";
+		}
+		
 		Board board = service.detail(bno);
 		model.addAttribute("detail",board);
 		
@@ -99,6 +164,17 @@ public class BoardController {
 	//게시글 수정
 	@RequestMapping(value="/board/update",method=RequestMethod.POST)
 	public String update(RedirectAttributes attr,HttpServletRequest request) {
+		//로그인이 안되어 있으면
+		if(request.getSession().getAttribute("userEmail") == null) {
+			//현재의 uri를 세션에 저장(/board/wirte)
+			//request.getRequestURI() 는 포트번호 이후의 경로를 반환 (String 형) ex./board/write
+			//request.getRequestURL() 은 http부터 끝까지의 경로를 반환 (StringBuffer 형) ex.http://localhost:8080/board/write 
+			String uri = request.getRequestURI();
+			request.getSession().setAttribute("uri",uri); 
+			//로그인 페이지가 없으므로 메인페이지로 리다이렉트
+			return "redirect:/";
+		}
+		
 		int bno = Integer.parseInt(request.getParameter("bno"));
 		boolean result = service.update(request);
 		//실패하면
